requirements.txt
jupyter
notebook
pandas
numpy
matplotlib
seaborn
wordcloud
nltk
scikit-learn
sentence-transformers
torch
tqdm
pinecone
psycopg2-binary
python-dotenv 
accelerate
huggingface_hub
